+++ .yalc/msupply-ble-service/package.json
@@ -135,5 +135,5 @@
   "dependencies": {
     "buffer": "^6.0.3"
   },
-  "yalcSig": "8040bc70b2b05776d4a1ad2df088683a"
+  "yalcSig": "9975ebcc878c53cca331550ba31d1b60"
 }
