+++ package.json
@@ -152,7 +152,8 @@
     "prettier": "^2.1.2",
     "react-require": "^1.0.1",
     "react-test-renderer": "17.0.1",
-    "redux-saga-test-plan": "^4.0.0-rc.3"
+    "redux-saga-test-plan": "^4.0.0-rc.3",
+    "yalc": "^1.0.0-pre.53"
   },
   "jest": {
     "preset": "react-native"
