+++ yalc.lock
@@ -2,7 +2,7 @@
   "version": "v1",
   "packages": {
     "msupply-ble-service": {
-      "signature": "8040bc70b2b05776d4a1ad2df088683a",
+      "signature": "9975ebcc878c53cca331550ba31d1b60",
       "file": true
     }
   }
